Bugfixes and first implementation of players.db reader
Bugfixes and added more Error Dialogs for better feedback and error tracking on the users side.
Added the first implementation of the Players.DB reader and added the feature to see al player positions and the basic information of the players inventory.
Added Encryption for the ErrorSender to prevent others to read Network trafic and get the information sent